modify base.html

+++ main project/BBS_Django/userprofile/models.py
@@ -16,7 +16,8 @@ def upload_to(instance, fielname):
 
     # 清除原来的头像图片
     remove_path = '/'.join([MEDIA_ROOT, instance.user.username])
-    os.remove(remove_path) 
+    if os.path.exists(remove_path):
+        os.remove(remove_path) 
     return path
 
 # 用户扩展信息
